backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pymysql
python-decouple
gunicorn